refactor: avoid duplication of matchCoordinate function
